+++ CHANGES.rst
@@ -46,6 +46,7 @@ Changelog
 - In `Migrate_To_4105._cleanFTWLabels`, be sure to keep old values in case still a `PersistentList` instead removing the annotation
 - In `Migrate_To_4105._removeBrokenAnnexes`, manage parent's modification date to keep old value because removing an annex
   will `notifyModifiedAndReindex` it's container
+- In `@@item_duplicate_form`, disable annexes if user does not have the permission to `Add annex/Add annexDecision` on future created item
 
 4.1.23.3 (2020-04-30)
 ---------------------